refactor code directory structure
