+++ libs/react-sketch-canvas/src/index.ts
@@ -1,3 +1,3 @@
 export * from './lib/Canvas';
 export * from './lib/ReactSketchCanvas';
-export * from './lib/typings';
+export * from './lib/types';
