+++ libs/react-sketch-canvas/src/lib/ReactSketchCanvas/ReactSketchCanvas.spec.tsx
