+++ libs/react-sketch-canvas/src/lib/ReactSketchCanvas/index.tsx
@@ -1,23 +1,23 @@
-import { produce } from "immer";
-import React from "react";
-import { Canvas } from "./Canvas";
-import { CanvasPath, ExportImageType, Point } from "./typings";
+import { produce } from 'immer';
+import React from 'react';
+import { Canvas } from '../Canvas';
+import { CanvasPath, ExportImageType, Point } from '../types';
 
 /* Default settings */
 
 const defaultProps = {
-  width: "100%",
-  height: "100%",
-  className: "",
-  canvasColor: "white",
-  strokeColor: "red",
-  background: "",
+  width: '100%',
+  height: '100%',
+  className: '',
+  canvasColor: 'white',
+  strokeColor: 'red',
+  background: '',
   strokeWidth: 4,
   eraserWidth: 8,
-  allowOnlyPointerType: "all",
+  allowOnlyPointerType: 'all',
   style: {
-    border: "0.0625rem solid #9c9c9c",
-    borderRadius: "0.25rem",
+    border: '0.0625rem solid #9c9c9c',
+    borderRadius: '0.25rem',
   },
   // eslint-disable-next-line @typescript-eslint/no-empty-function
   onUpdate: (_: CanvasPath[]): void => {},
@@ -294,7 +294,7 @@ export class ReactSketchCanvas extends React.Component<
     const exportImage = this.svgCanvas.current?.exportImage;
 
     if (!exportImage) {
-      throw Error("Export function called before canvas loaded");
+      throw Error('Export function called before canvas loaded');
     } else {
       return exportImage(imageType);
     }
@@ -305,7 +305,7 @@ export class ReactSketchCanvas extends React.Component<
       const exportSvg = this.svgCanvas.current?.exportSvg;
 
       if (!exportSvg) {
-        reject(Error("Export function called before canvas loaded"));
+        reject(Error('Export function called before canvas loaded'));
       } else {
         exportSvg()
           .then((data) => {
