+++ libs/react-sketch-canvas/src/lib/types/export.type.ts
@@ -1,0 +1,1 @@
+export type ExportImageType = 'jpeg' | 'png';
