+++ libs/react-sketch-canvas/src/lib/types/index.ts
@@ -1,0 +1,3 @@
+export * from './canvasPath.type';
+export * from './export.type';
+export * from './point.type';
