+++ libs/react-sketch-canvas/src/lib/types/point.type.ts
@@ -1,0 +1,4 @@
+export interface Point {
+  readonly x: number;
+  readonly y: number;
+}
